pyproject.toml
[project]
name = "njupt-smartclass-downloader"
version = "0.1.0"
description = ""
authors = [{ name = "ArcticLampyrid", email = "ArcticLampyrid@outlook.com" }]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "numpy (>=2.2.6,<3.0.0)",
    "ddddocr (>=1.5.6,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pycryptodome (>=3.23.0,<4.0.0)",
    "opencv-python-headless (>=4.11.0,<5.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "reportlab (>=4.4.1,<5.0.0)",
    "textual (>=3.3.0,<4.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "lxml (>=5.4.0,<6.0.0)",
    "sanitize-filename (>=1.2.0,<2.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "njupt_smartclass_downloader", from = "src" }]

[tool.poetry.scripts]
njupt-smartclass-downloader = "njupt_smartclass_downloader.__main__:main"
